Rename tests AssertUniqueTextRector to tests AssertNoUniqueTextRector since AssertUniqueText is now in the configurable rule
