+++ tests/src/Rector/Deprecation/AssertNoUniqueTextRector/config/configured_rule.php
