+++ tests/src/Rector/Deprecation/AssertNoUniqueTextRector/fixture/assert_no_unique_text.php.inc
