+++ tests/src/Rector/Deprecation/AssertNoUniqueTextRector/fixture/assert_no_unique_text_by_var.php.inc
