+++ tests/src/Rector/Deprecation/AssertNoUniqueTextRector/fixture/kerneltest_assert_no_unique_text.php.inc
